+++ ProductService/src/main/java/com/wolfhack/cloud/product/repository/CpuRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CpuRepository extends MongoRepository<Cpu, Double> {
+public interface CpuRepository extends MongoRepository<Cpu, Long> {
 
     @Query("{ $text: { $search: ?0 }}")
     Page<Cpu> searchCpusByQuery(String query, Pageable pageable);
